make provisioning less chatty
- removed -x, we are sufficiently sure about the script now
- silence apt-get and mkcert, unless an error appears
- provide status messages throughout the script
- user-specific options can be added to daemon.json by placing
  daemon-extra.json into the project directory -- it's content will
  placed verbatim right into daemon.json on the host

+++ provision.sh
@@ -4,7 +4,7 @@
 # daemon from download.docker.com, set up TCP/TLS and generate a
 # dockercontext to be consumed by clients.
 
-set -xeuo pipefail
+set -euo pipefail
 
 # About this virtual machine
 ARCH=$(dpkg --print-architecture)
@@ -26,24 +26,40 @@ MKCERT_PATH=/usr/bin/mkcert
 
 # Silently install APT packages
 function apt_install () {
-    apt-get -q update -q
-    apt-get -y -qq install "$@"
+    DEBIAN_FRONTEND=noninteractive apt-get -q update -q > /dev/null
+    if ! DEBIAN_FRONTEND=noninteractive apt-get -y -qq install "$@" &>apt-install.log ; then
+        echo "apt-get install $@ failed:"
+        cat apt-install.log >&2
+        exit 1
+    fi
 }
 
+# Run mkcert without the cruft
+function run_mkcert() {
+    if $MKCERT_PATH $@ 2>mkcert-stderr.log; then
+        rm -f mkcert-stderr.log
+    else
+        cat mkcert-stderr.log >&2
+        exit 1
+    fi
+}
 
 # Install Docker daemon if necessary
 function ensure_docker () {
-    if ! which dockerd; then
-	# Prerequisites for docker installation
-	apt_install ca-certificates curl gnupg lsb-release
+    if ! command -v dockerd &>/dev/null ; then
+        # Prerequisites for docker installation
+        echo "Installing containerd prerequisites ..."
+        apt_install ca-certificates curl gnupg lsb-release
 
-	# Get Docker GPG key and add docker repo to APT
-	curl -fsSL ${DOCKER_GPG_URL} | gpg --dearmor -o ${DOCKER_GPG}
-	echo "deb [arch=${ARCH} signed-by=${DOCKER_GPG}] ${DOCKER_REPO} \
-	  $( lsb_release -cs) stable" > ${DOCKER_APT_SRCS}
+        # Get Docker GPG key and add docker repo to APT
+        echo "Adding docker repo ${DOCKER_REPO} ..."
+        curl -fsSL ${DOCKER_GPG_URL} | gpg --dearmor -o ${DOCKER_GPG}
+        echo "deb [arch=${ARCH} signed-by=${DOCKER_GPG}] ${DOCKER_REPO} \
+          $( lsb_release -cs) stable" > ${DOCKER_APT_SRCS}
 
-	# Install Docker
-	apt_install docker-ce docker-ce-cli containerd.io
+        # Install Docker
+        echo "Installing docker CLI & containerd ..."
+        apt_install docker-ce docker-ce-cli containerd.io
     fi
 }
 
@@ -51,16 +67,18 @@ function ensure_docker () {
 # Install mkcert if necessary; mkcert generates locally valid
 # certificates (https://github.com/FiloSottile/mkcert)
 function ensure_mkcert () {
-    if ! which mkcert; then
-	rm -f mkcert-v*-linux-$(ARCH)
-	curl -o ${MKCERT_PATH} -JLO ${MKCERT_URL}
-	chmod +x ${MKCERT_PATH}
+    if ! command -v mkcert &>/dev/null ; then
+        echo "Installing mkcert ..."
+        rm -f mkcert-v*-linux-$ARCH
+        curl -o ${MKCERT_PATH} -sJL ${MKCERT_URL}
+        chmod +x ${MKCERT_PATH}
     fi
 }
 
-
 function make_dockercontext () {
-    [ -f meta.json ] || cat > meta.json <<EOF
+    if [ ! -f meta.json ]; then
+        echo "Creating docker context ..."
+        cat > meta.json <<EOF
 {
   "Name": "docker-parallels",
   "Metadata": {"Description": "dockerhost"},
@@ -72,21 +90,30 @@ function make_dockercontext () {
   }
 }
 EOF
-    mkdir -p tls/docker
-    [ -f tls/docker/cert.pem ] || mkcert -client -cert-file tls/docker/cert.pem -key-file tls/docker/key.pem host
-    [ -f tls/docker/ca.pem ] || cp /root/.local/share/mkcert/rootCA.pem tls/docker/ca.pem
-    tar cvf ${HOSTSHARE}/docker-parallels.dockercontext meta.json tls
+        mkdir -p tls/docker
+        echo "Creating client certificate ..."
+        [ -f tls/docker/cert.pem ] || run_mkcert -client -cert-file tls/docker/cert.pem -key-file tls/docker/key.pem host
+        [ -f tls/docker/ca.pem ] || cp /root/.local/share/mkcert/rootCA.pem tls/docker/ca.pem
+        tar cf ${HOSTSHARE}/docker-parallels.dockercontext meta.json tls
+    fi
 }
 
 function configure_docker () {
-    [ -f server.pem ] || mkcert -cert-file server.pem -key-file server-key.pem $MYIP
-    echo "Updating /etc/docker/daemon.json"
-    cat > /etc/docker/daemon.json <<EOF
+    if [ ! -f server.pem ]; then
+        echo "Creating server certificate ..."
+        run_mkcert -cert-file server.pem -key-file server-key.pem $MYIP
+        echo "Updating /etc/docker/daemon.json"
+        DAEMON_EXTRA_JSON=""
+        if [ -f /vagrant/daemon-extra.json ]; then
+            DAEMON_EXTRA_JSON=$(cat /vagrant/daemon-extra.json)
+        fi
+
+        cat > /etc/docker/daemon.json <<EOF
 {
    "experimental": false,
    "builder": { "gc": { "enabled": true, "defaultKeepStorage": "20GB" } },
    "features": { "buildkit": true },
-   "registry-mirrors": ["https://dockerhub.contact.de"],
+   $DAEMON_EXTRA_JSON
    "tls": true,
    "tlsverify": true,
    "tlscacert": "/root/.local/share/mkcert/rootCA.pem",
@@ -96,19 +123,23 @@ function configure_docker () {
    "containerd": "/run/containerd/containerd.sock"
 }
 EOF
-
-    mkdir -p /etc/systemd/system/docker.service.d
-    cat > /etc/systemd/system/docker.service.d/override.conf <<EOF
+        echo "Updating & restarting docker.service ..."
+        mkdir -p /etc/systemd/system/docker.service.d
+        cat > /etc/systemd/system/docker.service.d/override.conf <<EOF
 [Service]
 ExecStart=
 ExecStart=/usr/bin/dockerd
 EOF
 
-    systemctl daemon-reload
-    systemctl restart docker
+        systemctl daemon-reload
+        systemctl restart docker
+    fi
 
     echo "Installing binfmt support..."
-    docker run --privileged --rm tonistiigi/binfmt --install all
+    chmod +x /vagrant/setup-binfmt.sh
+    cp /vagrant/binfmt-setup.service /etc/systemd/system/
+    systemctl enable binfmt-setup.service
+    systemctl start binfmt-setup.service
 }
 
 cd /home/vagrant
